requirements.txt
streamlit==1.32.0
spacy==3.7.1
pyspellchecker==0.8.1
